kernel: fix up play_sound syscall
we want to pass in a 32-bit data, not 8-bit

+++ kernel/include/drivers/audio/audio.h
@@ -1,3 +1,3 @@
 uint32_t audio_pwm_init(void);
-int32_t audio_pwm_write(uint8_t *buffer, uint32_t len, uint32_t repeat);
+int32_t audio_pwm_write(uint32_t *buffer, uint32_t len, uint32_t repeat);
 int32_t audio_beep(void);
